Remove multiplexing capability (#4)
Before this commit, we used github.com/xtaci/smux to let us multiplex
several abstract network sessions atop a TCP connection. This feature
added significant complexity, and I'm not sure we need it. The way
I understand it, our only use case right now is to make a TCP
endpoint on our laptops that lives as long as the SSH connection
that uses it.