fix initial connection logic and add keep alives (#7)
This change fixes the initial connection logic to return errors if the
initial websocket dial fails. This change also adds keep alives to
client connections, this allows having to wait for a TCP timeout which
makes it more robust when the client's network changes/roams.

A bug where lastActivity was not being updated during a live flow was
also fixed.